Update Landing page with proper structure
The key was the height: 100vh setting because I wanted to have the
footer always show at the bototm of the page, but you can't do
this without a height number when flex-direction is column. It
took some time to figure out that 100vh is a dynamic height
based on the viewport size.